+++ game/GatewayConnRegistry.h
@@ -14,7 +14,7 @@ public:
         std::string token;  // fence_token
         std::string session_id;
         uint64_t generation = 0;
-        int connection_id = 0;
+        uint64_t connection_id = 0;
         std::string gamelogic_instance_id;
         uint64_t map_instance_id = 0;
         uint64_t map_owner_epoch = 0;
@@ -25,17 +25,17 @@ public:
 
     static GatewayConnRegistry &Instance();
 
-    void Remember(int connection_id, Bind bind);
-    void Forget(int connection_id);
+    void Remember(uint64_t connection_id, Bind bind);
+    void Forget(uint64_t connection_id);
     bool FindBySession(const std::string &session_id, Bind *out);
-    bool FindByConnection(int connection_id, Bind *out);
+    bool FindByConnection(uint64_t connection_id, Bind *out);
     bool FindByPlayer(uint64_t player_id, Bind *out);
     bool SendBySession(const std::string &session_id, const std::string &frame);
 
 private:
     GatewayConnRegistry() = default;
     std::mutex mu_;
-    std::unordered_map<int, Bind> by_conn_;
-    std::unordered_map<std::string, int> session_to_conn_;
-    std::unordered_map<uint64_t, int> player_to_conn_;
+    std::unordered_map<uint64_t, Bind> by_conn_;
+    std::unordered_map<std::string, uint64_t> session_to_conn_;
+    std::unordered_map<uint64_t, uint64_t> player_to_conn_;
 };
